refactor: Add middlewares in a dir
