+++ lib/all.js
@@ -3,13 +3,13 @@
 module.exports = {
   BaseLogger: require('./base_logger'),
   HttpLogger: require('./http_logger'),
-  HttpLoggerForExpress: require('./http_logger_for_express'),
-  HttpLoggerForKoa: require('./http_logger_for_koa'),
+  HttpLoggerForExpress: require('./middleware/http_logger_for_express'),
+  HttpLoggerForKoa: require('./middleware/http_logger_for_koa'),
   HttpMessage: require('./http_message'),
   HttpRequestImpl: require('./http_request_impl'),
   HttpResponseImpl: require('./http_response_impl'),
   HttpRule: require('./http_rule'),
   HttpRules: require('./http_rules'),
   UsageLoggers: require('./usage_loggers'),
-  WriterWrapper: require('./logged_response_wrapper')
+  WriterWrapper: require('./logged_response_wrapper'),
 };
