+++ lib/middleware/http_logger_for_express.js
@@ -1,15 +1,15 @@
 // © 2016-2022 Resurface Labs Inc.
 
-const HttpLogger = require('./http_logger');
-const HttpMessage = require('./http_message');
-const WriterWrapper = require('./logged_response_wrapper');
+const HttpLogger = require('../http_logger');
+const HttpMessage = require('../http_message');
+const WriterWrapper = require('../logged_response_wrapper');
 
 /**
- * Koa middleware for HTTP usage logging.
+ * Express middleware for HTTP usage logging.
  */
-class HttpLoggerForKoa {
+class HttpLoggerForExpress {
   /**
-   * Add new logger to the specified Koa app.
+   * Add new logger to the specified Express app.
    */
   static add(app, options = {}) {
     app.use(this.build(options));
@@ -19,7 +19,7 @@ class HttpLoggerForKoa {
    * Builds and initializes logger as Express middleware.
    */
   static build(options = {}) {
-    const logger = new HttpLoggerForKoa(options);
+    const logger = new HttpLoggerForExpress(options);
     return logger.handle.bind(logger);
   }
 
@@ -44,33 +44,35 @@ class HttpLoggerForKoa {
   /**
    * Called when request/response is passed through middleware.
    */
-  async handle(ctx, next) {
+  handle(request, response, next) {
     if (this._logger.enabled) {
-      this.log(ctx);
+      this.log(request, response, next);
+    } else {
+      next();
     }
-    await next();
   }
 
   /**
    * Logs the request/response from within middleware.
    */
-  log(ctx) {
-    ctx.started = HttpLogger.hrmillis;
+  log(request, response, next) {
     const logger = this._logger;
-    const wrapped = new WriterWrapper(ctx.res);
+    const started = HttpLogger.hrmillis;
+    const wrapped = new WriterWrapper(response);
 
     // declare event handler
     function afterResponse() {
-      ctx.res.removeListener('finish', afterResponse);
-      ctx.response.statusCode = ctx.status;
+      response.removeListener('finish', afterResponse);
       const now = Date.now().toString();
-      const interval = (HttpLogger.hrmillis - ctx.started).toString();
-      return HttpMessage.send(logger, ctx.request, ctx.response, wrapped.logged(), undefined, now, interval);
+      const interval = (HttpLogger.hrmillis - started).toString();
+      return HttpMessage.send(logger, request, response, wrapped.logged(), undefined, now, interval);
     }
 
     // register event handler
-    ctx.res.on('finish', afterResponse);
+    response.on('finish', afterResponse);
+
+    next();
   }
 }
 
-module.exports = HttpLoggerForKoa;
+module.exports = HttpLoggerForExpress;
