+++ lib/middleware/http_logger_for_koa.js
@@ -1,15 +1,15 @@
 // © 2016-2022 Resurface Labs Inc.
 
-const HttpLogger = require('./http_logger');
-const HttpMessage = require('./http_message');
-const WriterWrapper = require('./logged_response_wrapper');
+const HttpLogger = require('../http_logger');
+const HttpMessage = require('../http_message');
+const WriterWrapper = require('../logged_response_wrapper');
 
 /**
- * Express middleware for HTTP usage logging.
+ * Koa middleware for HTTP usage logging.
  */
-class HttpLoggerForExpress {
+class HttpLoggerForKoa {
   /**
-   * Add new logger to the specified Express app.
+   * Add new logger to the specified Koa app.
    */
   static add(app, options = {}) {
     app.use(this.build(options));
@@ -19,7 +19,7 @@ class HttpLoggerForExpress {
    * Builds and initializes logger as Express middleware.
    */
   static build(options = {}) {
-    const logger = new HttpLoggerForExpress(options);
+    const logger = new HttpLoggerForKoa(options);
     return logger.handle.bind(logger);
   }
 
@@ -44,35 +44,33 @@ class HttpLoggerForExpress {
   /**
    * Called when request/response is passed through middleware.
    */
-  handle(request, response, next) {
+  async handle(ctx, next) {
     if (this._logger.enabled) {
-      this.log(request, response, next);
-    } else {
-      next();
+      this.log(ctx);
     }
+    await next();
   }
 
   /**
    * Logs the request/response from within middleware.
    */
-  log(request, response, next) {
+  log(ctx) {
+    ctx.started = HttpLogger.hrmillis;
     const logger = this._logger;
-    const started = HttpLogger.hrmillis; 
-    const wrapped = new WriterWrapper(response);
+    const wrapped = new WriterWrapper(ctx.res);
 
     // declare event handler
     function afterResponse() {
-      response.removeListener('finish', afterResponse);
+      ctx.res.removeListener('finish', afterResponse);
+      ctx.response.statusCode = ctx.status;
       const now = Date.now().toString();
-      const interval = (HttpLogger.hrmillis - started).toString();
-      return HttpMessage.send(logger, request, response, wrapped.logged(), undefined, now, interval);
+      const interval = (HttpLogger.hrmillis - ctx.started).toString();
+      return HttpMessage.send(logger, ctx.request, ctx.response, wrapped.logged(), undefined, now, interval);
     }
 
     // register event handler
-    response.on('finish', afterResponse);
-
-    next();
+    ctx.res.on('finish', afterResponse);
   }
 }
 
-module.exports = HttpLoggerForExpress;
+module.exports = HttpLoggerForKoa;
